+++ examples/about_account.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from perch import API
+from perchx import API
 
 
 async def main():